runtime: remove paranoia/redundancies
This is all redundant code handled elsewhere by udev and mkinitcpio.
It's also silly to require that the "btrfs" binary be available at
build time, but then also do a runtime check... wat. I suspect this is
leftover from the days when btrfsctl was still around, and the "btrfs"
binary was the new hotness.

Additionally, add a note in the hook help to mention that the edgecase
of not using udev in your initramfs will require the btrfs hook from the
btrfs-progs package.