runtime: use more reasonable logical constructs
this removes some questionable use of compound commands to use more
readable if/then statements. Generally speaking, the form:

    a && b && c

Is *not* the same as:

    if a; then
      b
      c
    fi

Where the latter form is what's intended, but not what's logically
expressed. While rare, the former case will have unwanted behavior when
the middle of the expression can (and does) fail.